backend/requirements1.txt
accelerate==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.20.2
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.2
fastjsonschema==2.21.2
filelock==3.20.0
fqdn==1.5.1
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kornia==0.8.2
kornia_rs==0.1.10
lark==1.3.1
leditspp==1.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
moondream==0.1.1
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
segment-anything==1.0
Send2Trash==1.8.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.22
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
zipp==3.23.0
bchlib